+++ src/main/java/com/neueda/portfolio_manager/repository/TradeBookRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TradeBookRepository extends JpaRepository<TradeBook, Long> {
+public interface TradeBookRepository extends JpaRepository<TradeBook, Integer> {
 }
